+++ includes/install-functions.php
@@ -1,0 +1,84 @@
+<?php
+/**
+ * This file contains PHP functions that can be reused to download and install
+ * assets.
+ **/
+
+require_once( 'helper-functions.php' );
+
+function __installAsset( $json , $version ) {
+ global $bundler_version, $__data, $__cache;
+
+ $json = json_decode( $json , TRUE );
+
+ $name = $json[ 'name' ];
+ $type = $json[ 'type' ];
+
+ // Check to see if the version asked for is in the json; else, fallback to
+ // default if exists; if not, throw error.
+ if ( ! isset( $json[ 'versions' ][ $version ] ) ) {
+  if ( ! isset( $json[ 'versions' ][ 'default' ] ) ) {
+   echo "BUNDLER ERROR: No version found and cannot fall back to 'default' version!'";
+   return FALSE;
+  } else {
+   $version = 'default';
+  }
+ }
+
+ $get     = $json[ $version ][ 'get-method' ];
+ $invoke  = $json[ $version ][ 'invoke' ];
+ $install = $json[ $version ][ 'install' ];
+
+ // Download the file(s).
+ foreach ( $json[ '$versions' ][ $version ][ 'files' ] as $url ) {
+  $file = __doDownload( $url );
+  // File not found on the internets... DIE.
+  if ( $file == '5' ) return FALSE;
+ }
+
+ if ( $get == 'zip' ) {
+  // Unzip the file into the cache directory, silently.
+  exec( "unzip -qo '$__cache/$file' -d '$__cache'" );
+ } else if ( $get == 'tgz' || $get == 'tar.gz' ) {
+  // Untar the file into the cache directory, silently.
+  exec( "tar xzf '$__cache/$file' -C '$__cache'");
+ }
+
+ // For now, any other methods should be taken care of in the install
+ // instructions (in the JSON).
+
+ // Make sure that the directory structure exists before we try to put anything there.
+ __makeTree( "$__data/assets/$type/$name/$version" );
+
+ // Follow the installation instructions provided in the JSON file.
+ // This probably should be changed to work in accordance with the
+ // multiple file downloads above. However; well, that's for another
+ // version. Right now, the installation procedures should just be
+ // written out fully with what's above.
+ if ( is_array( $install ) ) {
+  foreach ( $install as $i ) {
+   // Replace the strings in the INSTALL json with the proper values.
+   $i = str_replace( "__FILE__" , $file , $i );
+   $i = str_replace( "__DATA__" , "$__data/assets/$type/$name/$version", $i );
+   exec( "$i" );
+  }
+ }
+
+ // Make the invoke file.
+ file_put_contents( "$__data/assets/$type/$name/$version/invoke" , $invoke );
+
+ __delTree( $__cache ); // Cleanup by deleting the cache directory.
+
+}
+
+function __doDownload( $url ) {
+ global $__cache;
+ $file = pathinfo( parse_url( "$url", PHP_URL_PATH ) );
+ exec( "curl -sL '$url' > '$__cache/" . $file['basename'] . "'" );
+ if ( file_get_contents( "$__cache/$file" ) == 'Not Found' ) {
+  echo "BUNDLER ERROR: Bad URL";
+  return 5;
+ } else {
+  return $file['basename'];
+ }
+}
